+++ app/src/main/java/com/dogwood008/kotlinrxsample/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.dogwood008.kotlin_rx_sample
+package com.dogwood008.kotlinrxsample
 
 import android.arch.lifecycle.ViewModelProviders
 import android.support.v7.app.AppCompatActivity
