+++ app/src/test/java/com/dogwood008/kotlinrxsample/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.dogwood008.kotlin_rx_sample
+package com.dogwood008.kotlinrxsample
 
 import org.junit.Test
